app/test-requirements.txt
pytest
pytest-cov
pytest-xdist
pytest-mock
pytest-html
pytest-benchmark
pytest-asyncio
pytest-timeout
pytest-randomly
pytest-clarity
bandit[toml]
safety
mypy
black
isort
flake8
pylint
memory-profiler
psutil
line-profiler
coverage[toml]
codecov
sphinx
sphinx-rtd-theme
pre-commit
tox
pytest-datadir
pytest-fixtures
pytest-mock-resources
pytest-repeat
pytest-rerunfailures
PyQt5
pyqtgraph
matplotlib
numpy
scipy
psutil
opencv-python
requests
websocket-client
pandas
pillow
pyserial
pathlib2
configparser
colorlog
packaging
setuptools
wheel
bleak
pyshimmer